Refactor MainActivity.kt
