+++ app/src/main/java/com/mochire/tech/ui/authentication/AuthActivity.kt
@@ -6,7 +6,7 @@ import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.mochire.tech.MainActivity
+import com.mochire.tech.ui.main.MainActivity
 import com.mochire.tech.R
